fix(core): secure path resolution for ripgrep and commandSafety
